added prv checks to add and edit filter buttons UI

+++ client/js/reports-manager.js
@@ -2338,6 +2338,11 @@ class ReportsManagerFilters extends Map {
 			</form>
 		`;
 
+		if(!this.stage.selectedReport.editable) {
+			container.querySelector('button.add-filter-button').title = 'Not enough privileges';
+			container.querySelector('button.add-filter-button').disabled = true;
+		}
+
 		const tbody = container.querySelector('table tbody');
 
 		for(const filter of this.values())
@@ -2709,12 +2714,17 @@ class ReportsManagerFilter {
 			<td>${this.type}</td>
 			<td>${this.default_value || (isNaN(parseFloat(this.offset)) ? '' : this.offset)}</td>
 			<td>${datasetName}</td>
-			<td class="action green"><i class="far fa-edit"></i></td>
-			<td class="action red"><i class="far fa-trash-alt"></i></td>
+			<td class="action ${this.stage.selectedReport.editable ? 'green' : 'grey'}"><i class="far fa-edit"></i></td>
+			<td class="action ${this.stage.selectedReport.deletable ? 'red' : 'grey'}"><i class="far fa-trash-alt"></i></td>
 		`;
 
-		row.querySelector('.green').on('click', () => this.edit());
-		row.querySelector('.red').on('click', () => this.delete());
+		if(row.querySelector('.green')) {
+			row.querySelector('.green').on('click', () => this.edit());
+		}
+
+		if(row.querySelector('.red')) {
+			row.querySelector('.red').on('click', () => this.delete());
+		}
 
 		row.on('mouseenter', () => {
